refactor: isolate filesystem search service from IPC wiring
